services/ai-platform/evaluation/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.27.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
pyyaml>=6.0
rich>=13.0.0
jinja2>=3.1.0
pytest>=8.0.0
